most of grid traversal complete, fly still bugged

+++ assignment02/worldgen.h
@@ -23,8 +23,8 @@
 
 typedef struct Location 
 {
-    uint8_t x;
-    uint8_t y;
+    int x;
+    int y;
 } Location_t;
 
 typedef struct Gates
